# Internship_API/src/main/resources/openapi/api.yaml
openapi: 3.0.1
info:
  title: Internship_API
  description: This is the Internship API documentation. My goal is to learn best practices about REST API Design and Architecture.
  version: 1.0.0
servers:
  - url: http://localhost:8081/v2
paths:
...
  /route/getAmountForReport:
    post:
      tags:
        - Route
      summary: Get report amount
      operationId: getAmountForReport
      security:
        - internship_api_security_scheme: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneralReportRequest"
      responses:
        "200":
          description: Report values
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: number
                  format: double
...
  /rent/get:
    get:
      tags:
        - Rent
      summary: Get all rents
      operationId: getRents
      security:
        - internship_api_security_scheme:
            - read_access
      parameters:
        - name: search
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/RentSearchObject"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/List"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/RentDTO"
  /rent/getAmountForReport:
    post:
      tags:
        - Rent
      summary: Get report amount
      operationId: getAmountForReportRent
      security:
        - internship_api_security_scheme: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneralReportRequest"
      responses:
        "200":
          description: Report values
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: number
                  format: double
  /rent/getById/{id}:
    get:
      tags:
        - Rent
      summary: Get rent by ID
      operationId: getRentById
      security:
        - internship_api_security_scheme:
            - read_access
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RentDTO"
        "404":
          description: The Rent with the specified ID does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  statusCode:
                    type: integer
  /rent/save:
    post:
      tags:
        - Rent
      summary: Create new rent
      operationId: saveRent
      security:
        - internship_api_security_scheme:
            - write_access
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RentInsertRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RentDTO"
  /rent/checkAvailability/{id}:
    post:
      tags:
        - Rent
      summary: Check rent availability for vehicle
      operationId: checkRentAvailability
      security:
        - internship_api_security_scheme:
            - read_access
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RentAvailabilityRequest"
      responses:
        "200":
          description: Availability status
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: boolean
        "404":
          description: The Rent with the specified ID does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  statusCode:
                    type: integer
  /rent/update/{id}:
    put:
      tags:
        - Rent
      summary: Update rent
      operationId: updateRent
      security:
        - internship_api_security_scheme:
            - write_access
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RentUpdateRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RentDTO"
        "404":
          description: The Rent with the specified ID does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  statusCode:
                    type: integer
  /rent/updateActive/{id}:
    put:
      tags:
        - Rent
      summary: Mark rent as active
      operationId: updateRentActive
      security:
        - internship_api_security_scheme:
            - write_access
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated to active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RentDTO"
        "404":
          description: The Rent with the specified ID does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  statusCode:
                    type: integer
  /rent/updateFinish/{id}:
    put:
      tags:
        - Rent
      summary: Finish rent
      operationId: updateRentFinish
      security:
        - internship_api_security_scheme:
            - write_access
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rent finished
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RentDTO"
        "404":
          description: The Rent with the specified ID does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  statusCode:
                    type: integer
  /rent/updatePayment/{id}:
    put:
      tags:
        - Rent
      summary: Mark rent as paid
      operationId: updateRentPayment
      security:
        - internship_api_security_scheme:
            - write_access
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RentDTO"
        "404":
          description: The Rent with the specified ID does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  statusCode:
                    type: integer
  /rent/delete/{id}:
    delete:
      tags:
        - Rent
      summary: Delete rent
      operationId: deleteRent
      security:
        - internship_api_security_scheme:
            - write_access
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RentDTO"
        "404":
          description: The Rent with the specified ID does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  statusCode:
                    type: integer
  #Notification paths
...
  /review/get:
    get:
      tags:
        - Review
      summary: Get all reviews
      operationId: getReviews
      security:
        - internship_api_security_scheme:
            - read_access
      parameters:
        - name: search
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ReviewSearchObject"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/List"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/ReviewDTO"
  /review/getAverageReview:
    get:
      tags:
        - Review
      summary: Get all reviews
      operationId: getAverageReview
      security:
        - internship_api_security_scheme:
            - read_access
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientReviewAverageDTO"
  /review/getById/{id}:
    get:
      tags:
        - Review
      summary: Get review by ID
      operationId: getReviewById
      security:
        - internship_api_security_scheme:
            - read_access
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewDTO"
        "404":
          description: The Review with the specified ID does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  statusCode:
                    type: integer
  /review/getDriversForReport:
    post:
      tags:
        - Review
      summary: Get drivers for review report
      operationId: getDriversForReport
      security:
        - internship_api_security_scheme:
            - read_access
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneralReportRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /review/save:
    post:
      tags:
        - Review
      summary: Create new review
      operationId: saveReview
      security:
        - internship_api_security_scheme:
            - write_access
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInsertRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewDTO"
  /review/update/{id}:
    put:
      tags:
        - Review
      summary: Update review
      operationId: updateReview
      security:
        - internship_api_security_scheme:
            - write_access
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewUpdateRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewDTO"
        "404":
          description: The Review with the specified ID does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  statusCode:
                    type: integer
  /review/delete/{id}:
    delete:
      tags:
        - Review
      summary: Delete review
      operationId: deleteReview
      security:
        - internship_api_security_scheme:
            - write_access
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewDTO"
        "404":
          description: The Review with the specified ID does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  statusCode:
                    type: integer
  #Statistics paths
...
components:
  securitySchemes:
    internship_api_security_scheme:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: http://keycloak:8080/realms/Internship_API/protocol/openid-connect/auth
          tokenUrl: http://keycloak:8080/realms/Internship_API/protocol/openid-connect/token
          scopes:
            read_access: read data
            write_access: modify data
  schemas:
    List:
      type: object
      properties:
        count:
          type: integer
    #User data objecets
    UserDTO:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        surname:
          type: string
        username:
          type: string
        email:
          type: string
        telephoneNumber:
          type: string
        gender:
          type: string
        registrationDate:
          type: string
          format: date-time
...
    ClientDTO:
      type: object
      properties:
        id:
          type: integer
        image:
          type: string
          format: byte
        user:
          $ref: "#/components/schemas/UserDTO"
...
    DriverDTO:
      type: object
      properties:
        id:
          type: integer
        numberOfClientsAmount:
          type: integer
        numberOfHoursAmount:
          type: integer
        user:
          $ref: "#/components/schemas/UserDTO"
...
    GeneralReportRequest:
      type: object
      required:
        - month
        - year
      properties:
        month:
          type: integer
          minimum: 1
          maximum: 12
        year:
          type: integer
          minimum: 2024
...
    VehicleDTO:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        available:
          type: boolean
        averageFuelConsumption:
          type: number
          format: double
        image:
          type: string
          format: byte
        price:
          type: number
          format: double
...
    RentDTO:
      type: object
      properties:
        id:
          type: integer
        rentDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        numberOfDays:
          type: integer
        fullPrice:
          type: number
          format: double
        paid:
          type: boolean
        status:
          type: string
        vehicle:
          $ref: "#/components/schemas/VehicleDTO"
        client:
          $ref: "#/components/schemas/ClientDTO"
    RentInsertRequest:
      type: object
      properties:
        rentDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        vehicle_id:
          type: integer
        client_id:
          type: integer
      required:
        - rentDate
        - endDate
    RentUpdateRequest:
      type: object
      properties:
        endDate:
          type: string
          format: date-time
        rentDate:
          type: string
          format: date-time
        vehicle_id:
          type: integer
    RentAvailabilityRequest:
      type: object
      properties:
        vehicle_id:
          type: integer
        rentDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
      required:
        - rentDate
        - endDate
    RentSearchObject:
      type: object
      properties:
        pageNumber:
          type: integer
        pageSize:
          type: integer
        status:
          type: string
        vehicle_id:
          type: integer
        client_id:
          type: integer
        user_id:
          type: integer
        rentDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
    # Notification data object
...
    ReviewDTO:
      type: object
      properties:
        id:
          type: integer
        value:
          type: integer
        description:
          type: string
        addingDate:
          type: string
          format: date-time
        client:
          $ref: "#/components/schemas/ClientDTO"
        driver:
          $ref: "#/components/schemas/DriverDTO"
    ReviewSearchObject:
      type: object
      properties:
        pageNumber:
          type: integer
        pageSize:
          type: integer
        reviewedName:
          type: string
        reviewsName:
          type: string
    ReviewInsertRequest:
      type: object
      properties:
        value:
          type: integer
          minimum: 1
          maximum: 5
        description:
          type: string
        reviewed_id:
          type: integer
        reviews_id:
          type: integer
      required:
        - value
        - description
        - reviewed_id
        - reviews_id
    ReviewUpdateRequest:
      type: object
      properties:
        value:
          type: integer
          minimum: 1
          maximum: 5
        description:
          type: string
      required:
        - value
        - description
    ClientReviewAverageDTO:
      type: object
      properties:
        client:
          type: string
        averageReview:
          type: number
          format: double  
    #Statistics data objects
...